Move functions out of main file
